# New parakeet-tdt-v3-0.6b models, ~50% faster 