--- a/Sources/FluidAudio/ASR/ANEOptimizer.swift
+++ b/Sources/FluidAudio/ASR/ANEOptimizer.swift
@@ -43,7 +43,7 @@ public enum ANEOptimizer {
 
     /// Configure optimal compute units for each model type
     public static func optimalComputeUnits(for modelType: ModelType) -> MLComputeUnits {
-        // Testing shows CPU+ANE is fastest for all models, including melspectrogram
+        // Testing shows CPU+ANE is fastest for all models, including fused mel encoder
         return .cpuAndNeuralEngine
     }
 
@@ -131,8 +131,7 @@ public enum ANEOptimizer {
 
     /// Model type enumeration for compute unit selection
     public enum ModelType {
-        case melSpectrogram
-        case encoder
+        case melEncoder
         case decoder
         case joint
     }