--- a/Sources/FluidAudio/ASR/AsrModels.swift
+++ b/Sources/FluidAudio/ASR/AsrModels.swift
@@ -8,8 +8,7 @@ public struct AsrModels: Sendable {
     /// Required model names for ASR
     public static let requiredModelNames = ModelNames.ASR.requiredModels
 
-    public let melspectrogram: MLModel
-    public let encoder: MLModel
+    public let melEncoder: MLModel
     public let decoder: MLModel
     public let joint: MLModel
     public let configuration: MLModelConfiguration
@@ -18,15 +17,13 @@ public struct AsrModels: Sendable {
     private static let logger = AppLogger(category: "AsrModels")
 
     public init(
-        melspectrogram: MLModel,
-        encoder: MLModel,
+        melEncoder: MLModel,
         decoder: MLModel,
         joint: MLModel,
         configuration: MLModelConfiguration,
         vocabulary: [Int: String]
     ) {
-        self.melspectrogram = melspectrogram
-        self.encoder = encoder
+        self.melEncoder = melEncoder
         self.decoder = decoder
         self.joint = joint
         self.configuration = configuration
@@ -69,8 +66,7 @@ extension AsrModels {
 
         // Load each model with its optimal compute unit configuration
         let modelConfigs: [(name: String, modelType: ANEOptimizer.ModelType)] = [
-            (Names.melspectrogramFile, .melSpectrogram),
-            (Names.encoderFile, .encoder),
+            (Names.melEncoderFile, .melEncoder),
             (Names.decoderFile, .decoder),
             (Names.jointFile, .joint),
         ]
@@ -93,8 +89,7 @@ extension AsrModels {
             }
         }
 
-        guard let melModel = loadedModels[Names.melspectrogramFile],
-            let encoderModel = loadedModels[Names.encoderFile],
+        guard let melEncoderModel = loadedModels[Names.melEncoderFile],
             let decoderModel = loadedModels[Names.decoderFile],
             let jointModel = loadedModels[Names.jointFile]
         else {
@@ -102,8 +97,7 @@ extension AsrModels {
         }
 
         let asrModels = AsrModels(
-            melspectrogram: melModel,
-            encoder: encoderModel,
+            melEncoder: melEncoderModel,
             decoder: decoderModel,
             joint: jointModel,
             configuration: config,
@@ -288,8 +282,7 @@ extension AsrModels {
         // The models will be downloaded to parentDir/parakeet-tdt-0.6b-v3-coreml/
         // by DownloadUtils.loadModels, so we don't need to download separately
         let modelNames = [
-            Names.melspectrogramFile,
-            Names.encoderFile,
+            Names.melEncoderFile,
             Names.decoderFile,
             Names.jointFile,
         ]
@@ -317,8 +310,7 @@ extension AsrModels {
     public static func modelsExist(at directory: URL) -> Bool {
         let fileManager = FileManager.default
         let modelFiles = [
-            Names.melspectrogramFile,
-            Names.encoderFile,
+            Names.melEncoderFile,
             Names.decoderFile,
             Names.jointFile,
         ]