… AGENTS.md
… Documentation/Benchmarks.md
… Sources/FluidAudio/ASR/TDT/TdtDecoderState.swift
… Sources/FluidAudio/Shared/ASRConstants.swift
… Sources/FluidAudioCLI/Commands/ASR/AsrBenchmark.swift
… Sources/FluidAudioCLI/Commands/ASR/FleursBenchmark.swift
… Sources/FluidAudioCLI/Commands/ASR/TranscribeCommand.swift
… Sources/FluidAudioCLI/Commands/DiarizationBenchmark.swift
… Sources/FluidAudioCLI/Commands/DownloadCommand.swift
… Sources/FluidAudioCLI/Commands/MultiStreamCommand.swift
… Sources/FluidAudioCLI/DatasetParsers/DatasetDownloader.swift
… Sources/FluidAudioCLI/Utils/ResultsFormatter.swift
… Sources/FluidAudioCLI/Utils/StreamingUI.swift
… Tests/FluidAudioTests/ANEOptimizerTests.swift
… Tests/FluidAudioTests/AsrManagerTests.swift
… Tests/FluidAudioTests/AsrModelsTests.swift
… Tests/FluidAudioTests/AsrTranscriptionTests.swift
… Tests/FluidAudioTests/DecoderStateTests.swift
… Tests/FluidAudioTests/TdtDecoderStateV3Tests.swift
… Tests/FluidAudioTests/TdtDecoderTests.swift