### Why is this change needed?
<!-- Explain the motivation for this change. What problem does it solve? -->

See [mobius](https://github.com/FluidInference/mobius) for details abuot the steps but TLDR is that
- We combined the preprocessor and the encoder into MelEncoder since they were ran contigously anyways
- We combined the Joint and some of the manual duration, proabbiltiy calculation into the Joint model into JointDecisions, because its also contiguous 
- Removed the transpose in the encoder, it was an optimization simply for the sake of working around our problem of not having an efficient way of using the encoder outputs 

We could probably remove the state dependency and just chunk on silences and batch the inference to speed it up even more, but this is decent enough for now IMO 

Note that this does use a bigger encoder, we found that the smaller 6bit one is ~100MB smaller but does impact the accuracy noticeably so going back to int8 


Before
```text
[21:29:28.718] [INFO] [FLEURSBenchmark] FLEURS BENCHMARK SUMMARY
[21:29:28.718] [INFO] [FLEURSBenchmark] ================================================================================
[21:29:28.718] [INFO] [FLEURSBenchmark]
[21:29:28.718] [INFO] [FLEURSBenchmark] Language                  | WER%   | CER%   | RTFx    | Duration | Processed | Skipped
[21:29:28.718] [INFO] [FLEURSBenchmark] -----------------------------------------------------------------------------------------
[21:29:28.718] [INFO] [FLEURSBenchmark] English (US)              | 5.7    | 2.8    | 125.9   | 3442.9s  | 350       | -
[21:29:28.718] [INFO] [FLEURSBenchmark] French (France)           | 5.8    | 2.4    | 122.5   | 560.8s   | 52        | 298
[21:29:28.718] [INFO] [FLEURSBenchmark] German (Germany)          | 3.1    | 1.2    | 140.4   | 62.1s    | 5         | -
[21:29:28.718] [INFO] [FLEURSBenchmark] Italian (Italy)           | 4.3    | 2.0    | 136.8   | 743.3s   | 50        | -
[21:29:28.718] [INFO] [FLEURSBenchmark] Russian (Russia)          | 7.7    | 2.8    | 122.5   | 621.2s   | 50        | -
[21:29:28.718] [INFO] [FLEURSBenchmark] Spanish (Spain)           | 6.4    | 3.0    | 139.4   | 586.9s   | 50        | -
[21:29:28.718] [INFO] [FLEURSBenchmark] Ukrainian (Ukraine)       | 6.5    | 1.9    | 121.9   | 528.2s   | 50        | -
[21:29:28.718] [INFO] [FLEURSBenchmark] -----------------------------------------------------------------------------------------
[21:29:28.718] [INFO] [FLEURSBenchmark] AVERAGE                   | 5.6    | 2.3    | 129.9   | 6545.5s  | 607       | 298
[21:29:28.718] [WARN] [FLEURSBenchmark]
```


After 
```text
[22:00:56.652] [INFO] [FLEURSBenchmark] ================================================================================
[22:00:56.652] [INFO] [FLEURSBenchmark] FLEURS BENCHMARK SUMMARY
[22:00:56.652] [INFO] [FLEURSBenchmark] ================================================================================
[22:00:56.652] [INFO] [FLEURSBenchmark]
[22:00:56.652] [INFO] [FLEURSBenchmark] Language                  | WER%   | CER%   | RTFx    | Duration | Processed | Skipped
[22:00:56.652] [INFO] [FLEURSBenchmark] -----------------------------------------------------------------------------------------
[22:00:56.652] [INFO] [FLEURSBenchmark] English (US)              | 5.8    | 2.9    | 189.4   | 3442.9s  | 350       | -
[22:00:56.652] [INFO] [FLEURSBenchmark] French (France)           | 8.8    | 3.8    | 181.3   | 560.8s   | 52        | 298
[22:00:56.652] [INFO] [FLEURSBenchmark] German (Germany)          | 4.2    | 1.2    | 211.2   | 62.1s    | 5         | -
[22:00:56.652] [INFO] [FLEURSBenchmark] Italian (Italy)           | 2.8    | 1.0    | 206.6   | 743.3s   | 50        | -
[22:00:56.652] [INFO] [FLEURSBenchmark] Russian (Russia)          | 7.0    | 2.3    | 185.3   | 621.2s   | 50        | -
[22:00:56.652] [INFO] [FLEURSBenchmark] Spanish (Spain)           | 4.0    | 1.8    | 207.9   | 586.9s   | 50        | -
[22:00:56.652] [INFO] [FLEURSBenchmark] Ukrainian (Ukraine)       | 7.2    | 2.1    | 182.8   | 528.2s   | 50        | -
[22:00:56.652] [INFO] [FLEURSBenchmark] -----------------------------------------------------------------------------------------
[22:00:56.652] [INFO] [FLEURSBenchmark] AVERAGE                   | 5.7    | 2.2    | 194.9   | 6545.5s  | 607       | 298
```

See Benchmarks.md for final comparison